MFC r307584 (partially)
307584 added the __min_size macro and replaced some static array sizes with
__min_size.  However, stable/10 doesn't yet have any static array sizes, so
all this MFC does is add the __min_size macro.

Original commit message:

Fix C++ includability of crypto headers with static array sizes

C99 allows array function parameters to use the static keyword for their
sizes. This tells the compiler that the parameter will have at least the
specified size, and calling code will fail to compile if that guarantee is
not met. However, this syntax is not legal in C++.

This commit reverts r300824, which worked around the problem for
sys/sys/md5.h only, and introduces a new macro: min_size(). min_size(x) can
be used in headers as a static array size, but will still compile in C++
mode.


git-svn-id: https://svn.freebsd.org/base/stable/10@309024 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f