MFC r308943
For RTT calculations mid-session, we explicitly ignore ACKs with tsecr of 0 as
many borken middle-boxes tend to do that. But during 3whs, in syncache_expand(),
we don't do that which causes us to send a RST to such a client. Relax this
constraint by only using tsecr to compare against timestamp that we sent when it
is not 0. As a result, we'd now accept the final ACK of 3whs with tsecr of 0.

Sponsored by:	Limelight Networks


git-svn-id: https://svn.freebsd.org/base/stable/11@309548 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f