MFC r305855, r306297, r306300, r306312-r306313
When MAKEOBJDIRPREFIX points to a case-insensitive file system, the
build can break when different source files create the same object
files (case-insensitivity speaking).  This is the case for object
files compiled with -fpic and shared libraries. The former uses
an extension of ".So", and the latter an extension ".so".  Rename
shared object files from *.So to *.pico to match what NetBSD does.

Also:
 o  Compile _Exit.c as C99_Exit.c, as it conflicts with _exit.s
 o  Add entry to UPDATING
 o  Document .pico extension


git-svn-id: https://svn.freebsd.org/base/stable/11@309843 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f