make smbus_get_addr() an ivar accessor
This is a direct commit to this branch.

smbus_get_addr() was declared as a funciton prototype, but the function
was never defined.  read_ivar and write_ivar methods were unused.

This was exposed after jedec_ts driver was imported to this branch
as that driver turned out to be the first user of smbus_get_addr().

This change is a small subset of a larger change in r281985.

Reported by:	jhb
Pointyhat to:	avg


git-svn-id: https://svn.freebsd.org/base/stable/10@310062 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f