MFC r309397:
Fix the handling of  TCP FIN-segments in the CLOSED state

When a TCP segment with the FIN bit set was received in the CLOSED state,
a TCP RST-ACK-segment is sent. When computing SEG.ACK for this, the
FIN counts as one byte. This accounting was missing and is fixed by this
patch.

Sponsored by:	Netflix, Inc.


git-svn-id: https://svn.freebsd.org/base/stable/11@310213 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f