MFC r306376:
Add a sysctl to control the interrupt pacing on AM335x integrated switch.
The hardware can be set to limit the number of interrupts from 2 to 63
interrupts per ms.

To keep the compatibility with the TI documentation the sysctl take the
interval between the interrupts pulses: 16~500 us.

Sponsored by:	Rubicon Communications, LLC (Netgate)


git-svn-id: https://svn.freebsd.org/base/stable/11@310857 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f